+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0.2
+
+* Reduced default long press duration from 1000ms to 700ms.
+* Added [onSecondaryTapUp] callback, allowing to intercept secondary tap event (such as mouse right click).
+* Added [TapDownDetails] parameter to [onLongPress] callback, making it possible to retrieve the local and global positions of the long press event.
+
+## 1.0.1
+
+* Added `onTapUp` callback, allowing to retrieve the `TapUpDetails` associated to the tap event.
+
 ## 1.0.0
 
 * Initial release